Add a check in PageView for the buffer mask  - fixes #5659  - prevents a crash when the mask is not set

+++ src/core/control/tools/PdfElemSelection.h
@@ -96,7 +96,7 @@ private:
     /// The PDF selection tool used for the selection.
     ToolType toolType;
 
-    long unsigned int selectionPageNr = npos;
+    size_t selectionPageNr = npos;
 
     /// The selection bounds. Note that this does not necessarily correspond to
     /// a rectangle--it may also indicate start and end positions of a linear
